add betaTester flag in workspace endpoint
